Fixed notebook tests and move those tests from widgets to jupyter
