--- tests/widgets/test_notebooks.py
@@ -1,25 +1,0 @@
-from testbook import testbook
-
-
-@testbook("notebooks/Tutorial.ipynb", execute=True)
-def test_tutorial(tb):
-    lpg = tb.ref("lpg")
-    client = tb.ref("client")
-
-    assert len(lpg.elements_by_id) > 0
-
-    assert len(lpg.elements_by_id) == len(client.elements_by_id)
-
-    assert not set(lpg.elements_by_id).symmetric_difference(set(client.elements_by_id))
-
-
-@testbook("dev_doc/Playbook Explorer Parts Test.ipynb", execute=True)
-def test_simple_parts_explorer(tb):
-    m0_interpretation = tb.ref("m0_interpretation")
-    assert m0_interpretation
-
-
-@testbook("dev_doc/SysML Model.ipynb", execute=True)
-def test_sysml_model(tb):
-    an_element = tb.ref("an_element")
-    assert an_element._id == "20511137-d751-4baf-ac4b-994a8e65ad3d"
